# contracts/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/rosters": {
      "post": {
        "operationId": "RostersController_createRoster",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create draft roster",
        "tags": [
          "roster"
        ]
      }
    },
    "/v1/rosters/current": {
      "get": {
        "operationId": "RostersController_current",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get currently active roster for authenticated user organization unit",
        "tags": [
          "roster"
        ]
      }
    },
    "/v1/rosters/{id}": {
      "get": {
        "operationId": "RostersController_byId",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get roster detail with shifts and assignments",
        "tags": [
          "roster"
        ]
      }
    },
    "/v1/rosters/{id}/shifts": {
      "post": {
        "operationId": "RostersController_createShift",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create shift in draft roster",
        "tags": [
          "roster"
        ]
      }
    },
    "/v1/rosters/{id}/shifts/{shiftId}": {
      "patch": {
        "operationId": "RostersController_updateShift",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "shiftId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update shift in draft roster",
        "tags": [
          "roster"
        ]
      },
      "delete": {
        "operationId": "RostersController_deleteShift",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "shiftId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete shift from draft roster",
        "tags": [
          "roster"
        ]
      }
    },
    "/v1/rosters/{id}/shifts/{shiftId}/assignments": {
      "post": {
        "operationId": "RostersController_assignShift",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "shiftId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Assign person to shift in draft roster",
        "tags": [
          "roster"
        ]
      }
    },
    "/v1/rosters/{id}/shifts/{shiftId}/assignments/{assignmentId}": {
      "delete": {
        "operationId": "RostersController_unassignShift",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "shiftId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assignmentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove person assignment from shift in draft roster",
        "tags": [
          "roster"
        ]
      }
    },
    "/v1/rosters/{id}/publish": {
      "post": {
        "operationId": "RostersController_publish",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Publish draft roster after min-staffing checks",
        "tags": [
          "roster"
        ]
      }
    },
    "/v1/rosters/{id}/plan-vs-actual": {
      "get": {
        "operationId": "RostersController_planVsActual",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Compute plan-vs-actual compliance for roster",
        "tags": [
          "roster"
        ]
      }
    },
...
    "/v1/integrations/webhooks/endpoints": {
      "post": {
        "operationId": "IntegrationsController_createEndpoint",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Register a webhook endpoint",
        "tags": [
          "integrations"
        ]
      },
      "get": {
        "operationId": "IntegrationsController_listEndpoints",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List webhook endpoints",
        "tags": [
          "integrations"
        ]
      }
    },
    "/v1/integrations/events/outbox": {
      "get": {
        "operationId": "IntegrationsController_outbox",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List outbox domain events",
        "tags": [
          "integrations"
        ]
      }
    },
    "/v1/integrations/webhooks/dispatch": {
      "post": {
        "operationId": "IntegrationsController_dispatch",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Dispatch pending outbox events to subscribed endpoints",
        "tags": [
          "integrations"
        ]
      }
    },
...
  },
  "info": {
    "title": "cueq API",
    "description": "Integrated time-tracking, absence-management, and shift-planning API for a German university (NRW / TV-L)",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "integration-token": {
        "type": "apiKey",
        "in": "header",
        "name": "x-integration-token",
        "description": "Deployment-injected shared token for terminal and HR machine endpoints"
      }
    },
...
  }
}